Moved things to Geometry.h

+++ Geometry.cpp
@@ -4,12 +4,12 @@
  *  Created on: Jan 3, 2023
  *      Author: Nathan
  */
-#include "Math.h"
+#include "Geometry.h"
 
 namespace priori{
 
 	Point::Point() : x(), y() {};
 
-	Point::Point(int x, int y) : x(x), y(y) {};
+	Point::Point(double x, double y) : x(x), y(y) {};
 
 }
